// src/models/task-check-item-vo.ts
export class TaskCheckItemVo {
  id: string;
  taskCheckId: string;
  itemTitle: string;
  itemSort: number;
  subItemId: string;
  subItemTitle: string;
  subItemSort: number;
  subItemItems: string;
  subItemRemark: string;
  subItemsChecked: string;
  subItemsImageUrls: string;
  remark: string;
}